+++ lib/plopHelpers.ts
@@ -1,8 +1,9 @@
 import type { NodePlopAPI } from "node-plop";
-import { capitalizeFirstLetter } from "../../../utils.js";
+import { capitalizeFirstLetter } from "./utils.js";
 
 const capitalizeHelperName = "capitalize";
 
+/** Добавляет кастомный хелпер, у стандартного проблемы с кириллицей */
 const addHelpers = (plop: NodePlopAPI) => {
   plop.setHelper(capitalizeHelperName, capitalizeFirstLetter);
 };
